event system tested in application.cpp with log system
Instance from all kind of event defined in game engine created. With the function of events IsInCategory(EventCategory) , tested to logging with compating all events' category with all event categories to provide which categories an event is in and is the event system works. If the event is in x different category (ex. input,keyboard) , then the toString() of event will be logged x times. With this, we will be also controlled the category system. When I run this if , if ,if ....isInCategory ..... for time of all events and categories, the output  is like in image which is in the path of ( testImages/firstVersionOfEventSystemTestOutput.png)